+++ fs/test/multiplatform/fs_api_test.dart
@@ -1,7 +1,3 @@
-// Copyright (c) 2015, <your name>. All rights reserved. Use of this source code
-// is governed by a BSD-style license that can be found in the LICENSE file.
-
-@Deprecated('do no used, move to fs_api_test in fs_shim')
 library;
 
 import 'package:fs_shim/fs_opfs_web.dart';